feat: add attributes stats for collections and create stats service
